@@ -4,6 +4,7 @@ buildscript {
     //  url 'https://oss.sonatype.org/content/repositories/snapshots/'
     //}
     mavenCentral()
+    jcenter()
   }
   dependencies {
     classpath 'com.android.tools.build:gradle:1.3.1'
@@ -13,7 +14,12 @@ buildscript {
       exclude module: 'guava'
     }
     classpath 'com.google.guava:guava:17.0'
+    classpath 'me.tatarka:gradle-retrolambda:3.2.3'
+    classpath 'me.tatarka.retrolambda.projectlombok:lombok.ast:0.2.3.a2'
   }
+
+  // Exclude the lombok version that the android plugin depends on.
+  configurations.classpath.exclude group: 'com.android.tools.external.lombok'
 }
 
 // Manifest version information!
@@ -25,6 +31,7 @@ def versionBuild = 0 // bump for dogfood builds, public betas, etc.
 apply plugin: 'com.android.application'
 apply plugin: 'com.neenbedankt.android-apt'
 apply plugin: 'spoon'
+apply plugin: 'me.tatarka.retrolambda'
 
 repositories {
   //maven {
@@ -94,8 +101,8 @@ android {
   }
 
   compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_7
-    targetCompatibility JavaVersion.VERSION_1_7
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
   }
 
   packagingOptions {
