@@ -338,11 +338,9 @@ public final class DebugView extends FrameLayout {
   private void setupMockBehaviorSection() {
     captureIntentsView.setEnabled(isMockMode);
     captureIntentsView.setChecked(captureIntents.get());
-    captureIntentsView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
-        Timber.d("Capture intents set to %s", b);
-        captureIntents.set(b);
-      }
+    captureIntentsView.setOnCheckedChangeListener((compoundButton, b) -> {
+      Timber.d("Capture intents set to %s", b);
+      captureIntents.set(b);
     });
 
     configureResponseSpinner(repositoriesResponseView, MockRepositoriesResponse.class);
@@ -399,47 +397,35 @@ public final class DebugView extends FrameLayout {
       }
     });
     // Ensure the animation speed value is always applied across app restarts.
-    post(new Runnable() {
-      @Override public void run() {
-        applyAnimationSpeed(animationSpeedValue);
-      }
-    });
+    post(() -> applyAnimationSpeed(animationSpeedValue));
 
     boolean gridEnabled = pixelGridEnabled.get();
     uiPixelGridView.setChecked(gridEnabled);
     uiPixelRatioView.setEnabled(gridEnabled);
-    uiPixelGridView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-        Timber.d("Setting pixel grid overlay enabled to " + isChecked);
-        pixelGridEnabled.set(isChecked);
-        uiPixelRatioView.setEnabled(isChecked);
-      }
+    uiPixelGridView.setOnCheckedChangeListener((buttonView, isChecked) -> {
+      Timber.d("Setting pixel grid overlay enabled to " + isChecked);
+      pixelGridEnabled.set(isChecked);
+      uiPixelRatioView.setEnabled(isChecked);
     });
 
     uiPixelRatioView.setChecked(pixelRatioEnabled.get());
-    uiPixelRatioView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-        Timber.d("Setting pixel scale overlay enabled to " + isChecked);
-        pixelRatioEnabled.set(isChecked);
-      }
+    uiPixelRatioView.setOnCheckedChangeListener((buttonView, isChecked) -> {
+      Timber.d("Setting pixel scale overlay enabled to " + isChecked);
+      pixelRatioEnabled.set(isChecked);
     });
 
     uiScalpelView.setChecked(scalpelEnabled.get());
     uiScalpelWireframeView.setEnabled(scalpelEnabled.get());
-    uiScalpelView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-        Timber.d("Setting scalpel interaction enabled to " + isChecked);
-        scalpelEnabled.set(isChecked);
-        uiScalpelWireframeView.setEnabled(isChecked);
-      }
+    uiScalpelView.setOnCheckedChangeListener((buttonView, isChecked) -> {
+      Timber.d("Setting scalpel interaction enabled to " + isChecked);
+      scalpelEnabled.set(isChecked);
+      uiScalpelWireframeView.setEnabled(isChecked);
     });
 
     uiScalpelWireframeView.setChecked(scalpelWireframeEnabled.get());
-    uiScalpelWireframeView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-        Timber.d("Setting scalpel wireframe enabled to " + isChecked);
-        scalpelWireframeEnabled.set(isChecked);
-      }
+    uiScalpelWireframeView.setOnCheckedChangeListener((buttonView, isChecked) -> {
+      Timber.d("Setting scalpel wireframe enabled to " + isChecked);
+      scalpelWireframeEnabled.set(isChecked);
     });
   }
 
@@ -471,12 +457,10 @@ public final class DebugView extends FrameLayout {
     boolean picassoDebuggingValue = picassoDebugging.get();
     picasso.setIndicatorsEnabled(picassoDebuggingValue);
     picassoIndicatorView.setChecked(picassoDebuggingValue);
-    picassoIndicatorView.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-      @Override public void onCheckedChanged(CompoundButton button, boolean isChecked) {
-        Timber.d("Setting Picasso debugging to " + isChecked);
-        picasso.setIndicatorsEnabled(isChecked);
-        picassoDebugging.set(isChecked);
-      }
+    picassoIndicatorView.setOnCheckedChangeListener((button, isChecked) -> {
+      Timber.d("Setting Picasso debugging to " + isChecked);
+      picasso.setIndicatorsEnabled(isChecked);
+      picassoDebugging.set(isChecked);
     });
 
     refreshPicassoStats();
@@ -568,44 +552,32 @@ public final class DebugView extends FrameLayout {
       hostView.setSelection(0, host.length()); // Pre-select it for editing.
 
       // Show the keyboard. Post this to the next frame when the dialog has been attached.
-      hostView.post(new Runnable() {
-        @Override public void run() {
-          Keyboards.showKeyboard(hostView);
-        }
-      });
+      hostView.post(() -> Keyboards.showKeyboard(hostView));
     }
 
     new AlertDialog.Builder(getContext()) //
         .setTitle("Set Network Proxy")
         .setView(view)
-        .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialog, int i) {
-            networkProxyView.setSelection(originalSelection);
-            dialog.cancel();
-          }
+        .setNegativeButton("Cancel", (dialog, i) -> {
+          networkProxyView.setSelection(originalSelection);
+          dialog.cancel();
         })
-        .setPositiveButton("Use", new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialog, int i) {
-            String in = hostView.getText().toString();
-            InetSocketAddress address = InetSocketAddressPreferenceAdapter.parse(in);
-            if (address != null) {
-              networkProxyAddress.set(address); // Persist across restarts.
-              proxyAdapter.notifyDataSetChanged(); // Tell the spinner to update.
-              networkProxyView.setSelection(ProxyAdapter.PROXY); // And show the proxy.
+        .setPositiveButton("Use", (dialog, i) -> {
+          String in = hostView.getText().toString();
+          InetSocketAddress address = InetSocketAddressPreferenceAdapter.parse(in);
+          if (address != null) {
+            networkProxyAddress.set(address); // Persist across restarts.
+            proxyAdapter.notifyDataSetChanged(); // Tell the spinner to update.
+            networkProxyView.setSelection(ProxyAdapter.PROXY); // And show the proxy.
 
-              Proxy proxy = InetSocketAddressPreferenceAdapter.createProxy(address);
-              client.setProxy(proxy);
-              apiClient.setProxy(proxy);
-            } else {
-              networkProxyView.setSelection(originalSelection);
-            }
-          }
-        })
-        .setOnCancelListener(new DialogInterface.OnCancelListener() {
-          @Override public void onCancel(DialogInterface dialogInterface) {
+            Proxy proxy = InetSocketAddressPreferenceAdapter.createProxy(address);
+            client.setProxy(proxy);
+            apiClient.setProxy(proxy);
+          } else {
             networkProxyView.setSelection(originalSelection);
           }
         })
+        .setOnCancelListener(dialogInterface -> networkProxyView.setSelection(originalSelection))
         .show();
   }
 
@@ -618,26 +590,20 @@ public final class DebugView extends FrameLayout {
     new AlertDialog.Builder(getContext()) //
         .setTitle("Set Network Endpoint")
         .setView(view)
-        .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialog, int i) {
-            endpointView.setSelection(originalSelection);
-            dialog.cancel();
-          }
+        .setNegativeButton("Cancel", (dialog, i) -> {
+          endpointView.setSelection(originalSelection);
+          dialog.cancel();
         })
-        .setPositiveButton("Use", new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialog, int i) {
+        .setPositiveButton("Use", (dialog, i) -> {
             String theUrl = url.getText().toString();
             if (!Strings.isBlank(theUrl)) {
               setEndpointAndRelaunch(theUrl);
             } else {
               endpointView.setSelection(originalSelection);
             }
-          }
         })
-        .setOnCancelListener(new DialogInterface.OnCancelListener() {
-          @Override public void onCancel(DialogInterface dialogInterface) {
+        .setOnCancelListener((dialogInterface) -> {
             endpointView.setSelection(originalSelection);
-          }
         })
         .show();
   }
