@@ -65,11 +65,7 @@ public final class DataModule {
   @Provides @Singleton Picasso providePicasso(Application app, OkHttpClient client) {
     return new Picasso.Builder(app)
         .downloader(new OkHttpDownloader(client))
-        .listener(new Picasso.Listener() {
-          @Override public void onImageLoadFailed(Picasso picasso, Uri uri, Exception e) {
-            Timber.e(e, "Failed to load image: %s", uri);
-          }
-        })
+        .listener((picasso, uri, e) -> Timber.e(e, "Failed to load image: %s", uri))
         .build();
   }
 
