@@ -7,13 +7,11 @@ import android.net.Uri;
 public interface IntentFactory {
   Intent createUrlIntent(String url);
 
-  IntentFactory REAL = new IntentFactory() {
-    @Override public Intent createUrlIntent(String url) {
-      Intent intent = new Intent(Intent.ACTION_VIEW);
-      intent.setData(Uri.parse(url));
-      intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+  IntentFactory REAL = url -> {
+    Intent intent = new Intent(Intent.ACTION_VIEW);
+    intent.setData(Uri.parse(url));
+    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
-      return intent;
-    }
+    return intent;
   };
 }
