@@ -11,9 +11,5 @@ public interface AppContainer {
   ViewGroup bind(Activity activity);
 
   /** An {@link AppContainer} which returns the normal activity content view. */
-  AppContainer DEFAULT = new AppContainer() {
-    @Override public ViewGroup bind(Activity activity) {
-      return findById(activity, android.R.id.content);
-    }
-  };
+  AppContainer DEFAULT = activity -> findById(activity, android.R.id.content);
 }
